+++ R/catmaid_annotations.R
@@ -20,8 +20,8 @@
 #'   For \code{catmaid_set_annotations_for_skeletons} a list containing 
 #'   information about the annotations that have just been added.
 
-#'   For \code{catmaid_remove_annotations_for_skeletons} a list containing 
-#'   information about the annotations that have just been removed.
+#' For \code{catmaid_remove_annotations_for_skeletons} a list containing
+#' information about the annotations that have just been removed.
 #' @export
 #' @examples
 #' \dontrun{
@@ -29,7 +29,9 @@
 #' catmaid_get_annotations_for_skeletons("name:ORN (left|right)")
 #' catmaid_get_annotations_for_skeletons("annotation:ORN PNs$")
 #' }
-#' @seealso \code{\link{catmaid_fetch}}, \code{\link{catmaid_skids}}
+#' @seealso \code{\link{catmaid_query_by_annotation}},
+#'   \code{\link{catmaid_get_meta_annotations}} \code{\link{catmaid_fetch}},
+#'   \code{\link{catmaid_skids}}
 catmaid_get_annotations_for_skeletons<-function(skids, pid=1, conn=NULL, ...) {
   skids=catmaid_skids(skids, conn = conn, pid=pid)
   post_data=list()
@@ -113,13 +115,17 @@ catmaid_remove_annotations_for_skeletons<-function(skids, annotations,
 
 # internal function to check for error state in return values
 # stops on error, otherwise returns input
-catmaid_error_check <- function(x){
+catmaid_error_check <- function(x, stoponerror=TRUE){
   err_fields=c("error","type", "detail")
-  if(length(x)==3 & all(names(x)%in%err_fields)){
-    # looks like an error response
-    stop(x$error, "\ttype: ", x$type)
+  if(length(x)>=3 & all(err_fields %in% names(x))){
+    # looks like an error response, let's get name of calling function
+    caller=sys.call(-1)[[1]]
+    if(stoponerror)
+      stop(caller, " of type: ", x$type, "\n  ", x$error, call. = FALSE)
+    else
+      return(TRUE)
   }
-  x
+  if(stoponerror) x else FALSE
 }
 
 #' Return the entity ids for one or more model ids
@@ -139,4 +145,159 @@ catmaid_entities_from_models <- function(skids, pid = 1, conn = NULL, ...) {
   path=sprintf("/%d/neurons/from-models", pid)
   unlist(catmaid_fetch(path, body=post_data, include_headers = F, 
                     simplifyVector = T, ...))
+}
+
+#' Get, query, set and remove CATMAID meta-annotations for annotations
+#'
+#' @description \code{catmaid_get_meta_annotations} Query by annotations to find
+#'   the meta-annotations that label them.
+#' @description \code{catmaid_query_meta_annotations} Query by meta-annotations
+#'   to find the annotations that they label.
+#' @description \code{catmaid_set_meta_annotations} Meta-annotate a group of
+#'   CATMAID annotations
+#' @description \code{catmaid_remove_meta_annotations} Remove meta-annotations
+#'   from annotations.
+#'
+#' @param annotations annotation ids designating which annotations to
+#'   meta-annotate. IDs can be found by calling
+#'   \code{catmaid_get_annotationlist}. If a character string is given, then IDs
+#'   will be found by calling \code{catmaid_get_annotationlist}.
+#' @param meta_annotations meta-annotation to add to query. Either a vector of
+#'   IDs or a character sting of meta-annotations can be given.
+#' @param with_annotations whether or not to return the other meta-annotations
+#'   of an annotation, when using \code{catmaid_query_meta_annotations}.
+#' @param force Whether to force the catmaid server to remove multiple
+#'   annotations (default \code{FALSE}, to provide some protection against
+#'   accidents).
+#' @inheritParams read.neuron.catmaid
+#' @seealso \code{\link{catmaid_query_by_annotation}} which has some overlap in
+#'   functionality, \code{\link{catmaid_get_annotations_for_skeletons}},
+#'   \code{\link{catmaid_skids}}, \code{\link{catmaid_get_annotationlist}}
+#' @export
+#' @name catmaid_meta_annotations
+#' @aliases catmaid_get_meta_annotations
+#' @examples
+#' \dontrun{
+#' ## Against FAFB CATMAID server
+#' catmaid_query_meta_annotations("ItoLee_Lineage")
+#'
+#' # note that this is similar to:
+#' catmaid_query_by_annotation("^ItoLee_Lineage$", type = 'annotation')
+#' }
+catmaid_get_meta_annotations <-function(annotations, pid=1, conn=NULL,...){
+  if(!possibly.numeric(annotations)){
+    a <- catmaid_get_annotationlist(pid=pid, conn=conn, ...)
+    annotations <- a$annotations[a$annotations$name%in%annotations,"id"]
+  }
+  if(!length(annotations)){
+    stop("Please give at least one valid annotation or annotation ID for your chosen CATMAID instance.")
+  }
+  post_data <-  list()
+  post_data[sprintf("object_ids[%d]", seq_along(annotations))] <- as.list(annotations)
+  path <- sprintf("/%d/annotations/query", pid)
+  res <- catmaid_fetch(path, body = post_data, include_headers = F,
+                       simplifyVector = T, conn = conn, ...)
+  invisible(catmaid_error_check(res))
+  res
+}
+
+#' @export
+#' @rdname catmaid_meta_annotations
+catmaid_query_meta_annotations <-function(meta_annotations, 
+                                      with_annotations = FALSE,
+                                      pid=1, conn=NULL,...){
+  if(!possibly.numeric(meta_annotations)){
+    a <- catmaid_get_annotationlist(pid=pid, conn=conn, ...)
+    meta_annotations <- a$annotations[a$annotations$name%in%meta_annotations,"id"]
+  }
+  if(!length(meta_annotations)){
+    stop("Please give at least one valid meta annotation or meta annotation ID for your chosen CATMAID instance.")
+  }
+  post_data <- list()
+  post_data[sprintf("annotated_with[%d]", seq_along(meta_annotations))] <- as.list(meta_annotations)
+  post_data["with_annotations"] <- with_annotations
+  post_data["types"] <- 'annotation'
+  path <- sprintf("/%d/annotations/query-targets", pid)
+  res <- catmaid_fetch(path, body = post_data, include_headers = F,
+                      simplifyVector = T, conn = conn, ...)
+  invisible(catmaid_error_check(res))
+  res$entities
+}
+
+#' @export
+#' @rdname catmaid_meta_annotations
+catmaid_set_meta_annotations<-function(meta_annotations,annotations,pid=1,conn=NULL,...){
+  if(!possibly.numeric(annotations)){
+    a <- catmaid_get_annotationlist(pid=pid, conn=conn, ...)
+    annotations <- a$annotations[a$annotations$name%in%annotations,"id"]
+  }
+  if(!length(annotations)){
+    stop("Please give at least one valid annotation or annotation ID for your chosen CATMAID instance.")
+  }
+  post_data <- list()
+  post_data[sprintf("entity_ids[%d]", seq_along(annotations))] <- as.list(annotations)
+  path <- sprintf("/%d/annotations/add", pid)
+  post_data[sprintf("annotations[%d]", seq_along(meta_annotations))] <- as.list(meta_annotations)
+  res <- catmaid_fetch(path, body = post_data, include_headers = F,
+                       simplifyVector = T, conn = conn, ...)
+  invisible(catmaid_error_check(res))
+}
+
+#' @export
+#' @rdname catmaid_meta_annotations
+catmaid_remove_meta_annotations <-function(annotations, 
+                                           meta_annotations,
+                                           force=FALSE, 
+                                           pid=1,
+                                           conn=NULL, ...) {
+  a <- 0
+  if(!possibly.numeric(annotations)){
+    a <- catmaid_get_annotationlist(pid=pid, conn=conn, ...)
+    annotations <- a$annotations[a$annotations$name%in%annotations,"id"]
+  }
+  if(!possibly.numeric(meta_annotations)){
+    if(a == 0){
+      a <- catmaid_get_annotationlist(pid=pid, conn=conn, ...) 
+    }
+    meta_annotations <- a$annotations[a$annotations$name%in%meta_annotations,"id"]
+  }
+  if(!length(meta_annotations) | !length(annotations)){
+    stop("Please give valid annotations or annotation IDs for your chosen CATMAID instance.")
+  }
+  post_data <-  list()
+  post_data[sprintf("entity_ids[%d]", seq_along(annotations))]=as.list(annotations)
+  if(length(meta_annotations)>1 && !force)
+    stop("You must set force=TRUE when removing multiple meta annotations")
+  post_data[sprintf("annotation_ids[%d]", seq_along(meta_annotations))]=as.list(meta_annotations)
+  path=sprintf("/%d/annotations/remove", pid)
+  res=catmaid_fetch(path, body=post_data, include_headers = F, 
+                    simplifyVector = T, ...)
+  invisible(catmaid_error_check(res))
+}
+
+# hidden
+possibly.numeric <- function(x) {
+  stopifnot(is.atomic(x) || is.list(x))
+  nNA <- sum(is.na(x))
+  nNA.new <- suppressWarnings(sum(is.na(as.numeric(x))))
+  nNA.new == nNA
+}
+
+#' Lock or unlock a CATMAID neuron reconstruction
+#'
+#' @description  Lock or unlock a CATMAID neuron reconstruction by adding or removing a 'locked' annotation to a set of skeleton IDs (skids). A locked neuron cannot be edited until it is unlocked.
+#' @inheritParams read.neuron.catmaid
+#' @export
+#' @rdname catmaid_lock_neurons
+catmaid_lock_neurons <- function(skids, pid = 1, conn = NULL, ...){
+  skids = catmaid_skids(skids, pid=pid,conn=conn,...)
+  catmaid_set_annotations_for_skeletons(skids, annotations = "locked", pid = pid,
+                                                 conn = conn, ...)
+}
+#' @export
+#' @rdname catmaid_lock_neurons
+catmaid_unlock_neurons <- function(skids, pid = 1, conn = NULL, ...){
+  skids = catmaid_skids(skids, pid=pid,conn=conn,...)
+  catmaid_remove_annotations_for_skeletons(skids, annotations = "locked", pid = pid,
+                                                    conn = conn, ...)
 }